+++ src/main/java/com/jamierf/mediamanager/parsing/rss/parsers/BitMeTVParser.java
@@ -11,6 +11,6 @@ public class BitMeTVParser extends RSSParser {
 	private static final String FEED_URL = "http://www.bitmetv.org/rss.php?uid=%d&passkey=%s";
 
 	public BitMeTVParser(HttpClientFactory clientFactory, ParserConfiguration config) throws MalformedURLException, ParserException {
-		super (clientFactory, String.format(FEED_URL, config.getLong("userId"), config.getString("passKey")));
+		super (clientFactory, String.format(FEED_URL, config.getInt("userId"), config.getString("passKey")));
 	}
 }
